requirements.txt
streamlit
keras
numpy
pillow
opencv-python-headless
streamlit-drawable-canvas